\name{EMaxDDGBM}
\alias{EMaxDDGBM}
\title{Expected Drawdown using Brownian Motion Assumptions}
\usage{
  EMaxDDGBM(R, digits = 4)
}
\arguments{
  \item{R}{an xts, vector, matrix, data frame, timeSeries
  or zoo object of asset returns}
}
\description{
  Works on the model specified by Maddon-Ismail
}
\author{
  R
}
\keyword{Assumptions}
\keyword{Brownian}
\keyword{Drawdown}
\keyword{Expected}
\keyword{Motion}
\keyword{Using}

